EFI bootloader fully working! ISO bootloader fixed! - Tons of configure.ac fixes - Tons of makefile fixes - Add bios and efi variations to all emulator launchers - Add assert macro to bootloader - Clear direction flag when returning from BIOS call - Add many files missing from project files - Change TEXT("example") macro to TEXT "example" - Drastically improve capability and robustness of BIOS disk read code - Support getting sector size from BIOS - Support 64-bit destination addresses - Automatic bounce buffering - Move getting ACPI and MPS data into bootloader, get same from EFI in EFI loader - Drastically improve bootloader page table manipulation code - Support getting I/O vectors for virtual address range - Drastically speed up and simply ELF loader code - Add inttypes.h macros such as PRIx64 - Add new memory types to physmem_data.h - Use the memory map when allocating memory in loader to avoid reserved ranges - Bootloader malloc improvements - Fix print_line to be more portable and not assume type sizes - Add char16_t variations of several string functions - Abstract keyboard input and screen output to implement EFI boot menu - Extend gensymtab.h to support extraction of symbols from PE files - Add kernel inline functions for setting all of the segment registers - Load every segment register in cpu_init_early - Avoid triple faults in more scenarios if an exception occurs in early boot - Move tons of entry.S assembly into C++ code - Add script for verifying that every file listed in project files exists - Add script for verifying that every source file is listed in a project file - Generate symbol files for EFI bootloader - Fix gcc patch to properly include all lines in LIBGCC config - Add a few missing "or exit" calls in gcc build script - Lots more!

+++ boot/mpentry.h
@@ -3,5 +3,5 @@
 #include "types.h"
 
 extern char const ap_entry[];
-extern size_t const ap_entry_size;
+extern uint32_t const ap_entry_size;
 extern uintptr_t ap_entry_vector;
